Move BasicContextProvider into project
